requirements.txt
flask==3.0.3
flask-cors==4.0.0
gunicorn==22.0.0
PyJWT==2.8.0
werkzeug==3.0.3
mercadopago==2.2.2
psycopg2-binary==2.9.9
